requirements.txt
asyncio
aiohttp
pandas
numpy
datetime
python-dateutil